Add ability to bypass certificate validation on HA Url (#48)
* Update media_lights_sync.py

I have a self-generated cert on my local HA instance which cannot be validated and results in an SSL issue. This adds a new param called `verify_cert`. It defaults to `True` when not specified in `apps.yaml`, but when set to `False`, it will skip validation of the SSL cert on the `ha_url` endpoint.

* Update README.md